requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
colorama==0.4.3
distlib==0.3.1
Django==3.0.7
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipenv==2020.8.13
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
wrapt==1.12.1
